requirements.txt
flask
gunicorn
groq
python-dotenv
markdown
openai